Clean up hover/active states in problem list
The hover state is now a much more subtle change, making
the blue "current" state more unique and pronounced. A
CSS transition should appear only between the normal and
hover states, but not others (which should respond quickly
to physical clicks).

Added an active state, in which the list item appears
pressed down.